% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locallevel_pop_props.R
\name{aggregate_indiv_categories}
\alias{aggregate_indiv_categories}
\title{Aggregate Individual Simulation Data into Categories}
\usage{
aggregate_indiv_categories(df = res.df1)
}
\arguments{
\item{df}{Individual level risk factor/demographic data for each la (list) from \code{\link{locallevel_pop_sim}}.}
}
\value{
area aggregated list
}
\description{
Aggregate Individual Simulation Data into Categories
}
\examples{

}
