% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locallevel_pop_props.R
\name{area_adjusted_agesex_probs}
\alias{area_adjusted_agesex_probs}
\title{Area_adjusted_agesex_probs}
\usage{
area_adjusted_agesex_probs(la.dat, agesex.dat, la.vals, agesex.vals, lanames,
  la, ages, sexes)
}
\arguments{
\item{la.dat}{Area (LA) level data}

\item{agesex.dat}{Age-sex data}

\item{la.vals}{Column name in la.dat of probability}

\item{agesex.vals}{Column name in agesex.dat}

\item{lanames}{Column name of LA names}

\item{la}{Single LA name to calculate result}

\item{ages}{Character vector of ages of interest (repeated for males and females)}

\item{sexes}{Character vector of sexes (repeated for each ageof interest)}
}
\value{
probs
}
\description{
Area_adjusted_agesex_probs
}
