% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinAllOutcomeData.R
\name{calcStats.CTADGUM_pred}
\alias{calcStats.CTADGUM_pred}
\title{Calculate Output Statistics of MRP Natsal Analysis}
\usage{
calcStats.CTADGUM_pred(CTADGUM_pred)
}
\arguments{
\item{CTADGUM_pred}{Combined model LA level predictions and census/administrative data}
}
\value{
CTADGUM_pred
}
\description{
Calculates the residuals, indirect standardised rates,
mean difference, binomial threshold exceedence and quadrant.
}
\seealso{
joinAllOutcomeData
}
