% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinAllOutcomeData.R
\name{calcStats.CTADGUM_pred_knnlog}
\alias{calcStats.CTADGUM_pred_knnlog}
\title{Calculate Output Statistics of knn and logistic Natsal Analysis}
\usage{
calcStats.CTADGUM_pred_knnlog(CTADGUM_pred)
}
\arguments{
\item{CTADGUM_pred}{}
}
\description{
Calculates the residuals, indirect standardised rates,
mean difference, binomial threshold exceedence and quadrant.
}
