% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning.R
\name{cleanDummyNames}
\alias{cleanDummyNames}
\title{Clean Dummy Names}
\usage{
cleanDummyNames(res)
}
\arguments{
\item{res}{}
}
\value{
names
}
\description{
clean names after expanding to
dummy variable format
}
\examples{

}
