% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_fns.R
\name{colNameReplace}
\alias{colNameReplace}
\title{Replace One Set of Column Names With Another}
\usage{
colNameReplace(array, name.before, name.after)
}
\arguments{
\item{name.after}{}
}
\value{
array
}
\description{
Replace One Set of Column Names With Another
}
