% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert2Groups.R
\name{convert2Groups}
\alias{convert2Groups}
\title{Generic Convert to Group}
\usage{
convert2Groups(ctsdata, groups, breaks)
}
\arguments{
\item{breaks}{}
}
\value{
character strings of categories
}
\description{
Generic Convert to Group
}
\examples{

}
