% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert2Groups.R
\name{convertAge2ageGroup}
\alias{convertAge2ageGroup}
\title{Convert Age to Age Group}
\usage{
convertAge2ageGroup(age)
}
\arguments{
\item{age}{(integer)}
}
\value{
character strings of categories
}
\description{
Convert Age to Age Group
}
\examples{

}
