% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert2Groups.R
\name{convertAge2ageGroup2}
\alias{convertAge2ageGroup2}
\title{Convert Age to Age Group Breaks 16-19, 20-21, 22-24}
\usage{
convertAge2ageGroup2(age)
}
\arguments{
\item{age}{(integer)}
}
\value{
character strings of categories
}
\description{
Convert Age to Age Group Breaks 16-19, 20-21, 22-24
}
\examples{

}
