% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Convert2Groups.R
\name{convertIncome2incomeGroup}
\alias{convertIncome2incomeGroup}
\title{Convert Income to Income Group}
\usage{
convertIncome2incomeGroup(income)
}
\arguments{
\item{income}{(integer)}
}
\value{
character strings of categories
}
\description{
Convert Income to Income Group
}
\examples{

}
