% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate-shift-fns.R
\name{covariateShift.glm}
\alias{covariateShift.glm}
\title{Covariate Shift GLM}
\usage{
covariateShift.glm(data, resla, riskfac, ssize = 10000)
}
\arguments{
\item{ssize}{}
}
\value{
data
}
\description{
alternative model-based approach:
 could fit a logistic regression to estimate the ratio of probabilities of each data set
 and then predict for (all) permutations
 http://blog.smola.org/post/4110255196/real-simple-covariate-shift-correction
}
