% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate-shift-fns.R
\name{covariateShift}
\alias{covariateShift}
\title{Covariate Shift}
\usage{
covariateShift(data, resla, riskfac, ssize = 10000)
}
\arguments{
\item{ssize}{}
}
\value{
data
}
\description{
Importance sampling approach
when different distributions for the
training and test data
}
