% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_covariates-man.R,
%   R/DATA_surveillance-man.R
\docType{data}
\name{CTADGUM.dat2011}
\alias{CTADGUM.dat2011}
\alias{CTADGUM.dat2011}
\title{Data
Urban-rural area proportions and classification.}
\description{
Major Urban
Large Urban
Other Urban
Large Market
Rural Town
Village
Dispersed

Older ages aren't available.
}
\references{
\url{data_blah.com}

\url{https://www.gov.uk/government/statistics/national-chlamydia-screening-programme-ncsp-data-tables}
}
\author{
My Name \email{blahblah@roxygen.org}

My Name \email{blahblah@roxygen.org}
}
\keyword{data}
