% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_surveillance-man.R
\docType{data}
\name{CTADGUM.dat2012}
\alias{CTADGUM.dat2012}
\title{Data
16 to 34 year old}
\description{
downloaded 13/5/16
CTAD Tables annual 2012 republished for publication_FINAL13062014
}
\references{
\url{https://www.gov.uk/government/statistics/national-chlamydia-screening-programme-ncsp-data-tables}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
