% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_surveillance-man.R
\docType{data}
\name{CTADGUM.dat2013}
\alias{CTADGUM.dat2013}
\title{Data
16 to 34 year old}
\description{
downloaded 11/5/16
CTAD Data Tables 2013 Annual data  for publication_FINAL130614
}
\references{
\url{https://www.gov.uk/government/statistics/national-chlamydia-screening-programme-ncsp-data-tables}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
