% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_fns.R
\name{DistnTable}
\alias{DistnTable}
\title{Distribution Table}
\usage{
DistnTable(data, riskfac)
}
\arguments{
\item{riskfac}{}
}
\value{
array
}
\description{
frequency table as dataframe
of discrete joint distribution
}
