% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_covariates-man.R
\docType{data}
\name{houseprice_earnings.dat}
\alias{houseprice_earnings.dat}
\title{Data
House price in relation to earnings index.}
\description{
Years: 1997-2013
Local Authority
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
