% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_covariates-man.R
\docType{data}
\name{IMD.dat2010}
\alias{IMD.dat2010}
\title{Data}
\description{
Index of Multiple Deprivation in 2010.
Local Authority
Income
Employment
Ranks
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
