% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joinAllOutcomeData.R
\name{joinAllOutcomeData}
\alias{joinAllOutcomeData}
\title{Join All Outcome Data}
\usage{
joinAllOutcomeData(pred, adj = 0.87, survData.filename = NA, ...)
}
\arguments{
\item{pred}{Prediction array. Ensure that the first column is the LA names.}

\item{adj}{Double counting adjustment <=1. From Sarah Woodall's paper.}

\item{survData.filename}{User supplied surveillance data}
}
\value{
CTADGUM_pred
}
\description{
Combines area level post-stratified predictions with surveillance data.
Take predictions from different models.
}
\details{
##TODO##
I've commented-out the logisitic knn predictions and used the MRP one instead
}
