% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn-fns.R
\name{knnDatamatrix}
\alias{knnDatamatrix}
\title{knnDatamatrix}
\usage{
knnDatamatrix(df, vars = "smokenow", c = 0.5)
}
\arguments{
\item{c}{age scaling}
}
\value{
data.matrix
}
\description{
create KNN input matrix
 comment-out unwanted variables
}
\details{
LA-level estimates to simulate with
 "sparsification" with dummy()
}
