% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning.R
\name{LAnameClean}
\alias{LAnameClean}
\title{LA name Clean}
\usage{
LAnameClean(datName)
}
\arguments{
\item{datName}{Vector of LA names}
}
\value{
datName Vector of cleaned LA names
}
\description{
Remove extra characters
and other consistencies
from Local Authority names
}
\examples{

}
