% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning.R
\name{levelsMatch}
\alias{levelsMatch}
\title{levelsMatch}
\usage{
levelsMatch(target, modify)
}
\arguments{
\item{modify}{}
}
\value{
factor
}
\description{
match-up number and order of levels
 in two vectors
}
\examples{

}
