% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locallevel_pop_props_class-TODO-.R
\name{locallevel_pop_props.class}
\alias{locallevel_pop_props.class}
\title{Estimate Local Level Population Proportions}
\usage{
locallevel_pop_props.class(input)
}
\value{
list by each LA
}
\description{
Combined LA level data for single risk factors and non-LA but multiple risk factor data sets.
Estimates a (partial-) joint distribution for each LA using the reference tables.
This is in contrast to \code{\link{locallevel_pop_sim}} which generates a sample of individuals
so is much slower to run.
Makes relative (to national average) LA-level prevalence adjustments.
}
\details{
Input data is in \code{data(pop_sim_input)} in \code{STIecoPredict} package.
}
\seealso{
\code{\link{locallevel_pop_sim}}
}
