% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locallevel_pop_sim.R
\name{locallevel_pop_sim}
\alias{locallevel_pop_sim}
\title{Local Level Population Simulation}
\usage{
locallevel_pop_sim(NATSAL.dat = NA, n = 100)
}
\arguments{
\item{NATSAL.dat}{NATSAL study data extract}

\item{n}{Number of individuals per LA to simulate}
}
\value{
list by each LA
}
\description{
Combined LA level data for single risk factors and non-LA but multiple risk factor data sets.
Simulate a sample, for each LA using the reference tables.
Makes relative (to national average) LA-level prevalence adjustments
}
\examples{

}
