% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps_fns.R
\name{map.tb}
\alias{map.tb}
\title{Choropleth Mapping for Natsal MRP Predictions}
\usage{
map.tb(map.data, g = NA, brks = NA, title = "", zoom = FALSE,
  value = "tb", file = FALSE)
}
\arguments{
\item{map.data}{Map object loaded by readOGR()}

\item{g}{Number of groups for values; max(g)=9}

\item{brks}{Given break points to group values (vector)}

\item{title}{Plot title (string)}

\item{zoom}{Whether to zoom-in (for UK) (logical)}

\item{value}{Variable name of colouring value (string)}

\item{file}{Output to file (logical)}
}
\value{
plot
}
\description{
Plot UK choropleth maps
 developed using LA and ETS data but
 can be applied more generally
 Note that the legend title is also used for the file name
}
