% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DATA_covariates-man.R
\docType{data}
\name{popLA}
\alias{popLA}
\title{Data
Local Authority population sizes, area and density.}
\description{
Data
Local Authority population sizes, area and density.
}
\references{
\url{data_blah.com}
}
\author{
My Name \email{blahblah@roxygen.org}
}
\keyword{data}
