% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_fns.R
\name{predict.lgstc}
\alias{predict.lgstc}
\title{Predict Logistic}
\usage{
\method{predict}{lgstc}(fit, data, ADD = F)
}
\arguments{
\item{ADD}{add to existing plot}
}
\description{
predict coverage from
logistic classifier model
}
