% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn-fns.R
\name{predictKNN}
\alias{predictKNN}
\title{predict K-Nearest Neighbour}
\usage{
predictKNN(fit, test = FALSE)
}
\arguments{
\item{test}{}
}
\value{
list
}
\description{
k-nerarest neighbours classifier prediction
NB only works for numeric covariates
careful when some of these are categorical imposing a (numerical) order
}
