% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classifier_fns.R
\name{scatterplotObsEst}
\alias{scatterplotObsEst}
\title{scatterplotObsEst}
\usage{
scatterplotObsEst(CTADGUM_pred, names, points)
}
\arguments{
\item{points}{}
}
\value{
plot
}
\description{
scatterplotObsEst
}
