% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTotalProbs.R
\name{SimInputPropData}
\alias{SimInputPropData}
\title{Check Simulated Input Proportions Data}
\usage{
SimInputPropData(input, formula = as.formula(cttestly ~ 1 + Sex + Age + drink
  + smoke + ethnic2 + student))
}
\arguments{
\item{input}{}
}
\value{
prop_props class object
}
\description{
Tests the integrity of the input data list for combining into the conditional joint data set.
}
