% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-cleaning.R
\name{toupper.fac}
\alias{toupper.fac}
\title{toupper.fac}
\usage{
toupper.fac(data, names)
}
\arguments{
\item{data}{dataframe}

\item{names}{factors (char)}
}
\value{
data dataframe
}
\description{
Coerce dataframe multiple factors to upper case
}
\details{
Column-wise operation
.
}
\examples{

}
